Rename to uptake error rate
