+++ checks/remotesettings/uptake_error_rate.py
